Add `block_while_running` option to `_ReadWriteThread`.
This way users don't need to understand the following distinctions:

* written to driver vs. output from card.
* `SDF_BUSY` vs. `SDF_RUNNING`.

The `subdevice.cancel` call in `_ReadWriteThread.block` is because the
`RUNNING` flag is cleared after output completes, but the `BUSY` flag
is not.  This may be a bug in Comedi.  I've asked on the mailing list,
but haven't heard back yet.